Rename project to Wurst v1.7 keybind converter
