--- Wurst v1.6 alt list converter/src/tk/wurst_client/v1_6_alt_list_converter/resources/options.json
@@ -1,16 +1,0 @@
-{
-  "__comment": [
-    "Copyright © 2015 | Alexander01998 | All rights reserved.",
-    "This Source Code Form is subject to the terms of the Mozilla Public",
-    "License, v. 2.0. If a copy of the MPL was not distributed with this",
-    "file, You can obtain one at http://mozilla.org/MPL/2.0/."
-  ],
-  "options": {
-    "alts": {
-      "enabled": true,
-      "label": "Alt list",
-      "oldfile": "alts.wurst",
-      "newfile": "alts.json"
-    }
-  }
-}
