+++ Wurst v1.7 keybind converter/src/tk/wurst_client/v1_7_keybind_converter/menu/DoneMenu.java
@@ -1,0 +1,27 @@
+/*
+ * Copyright © 2015 | Alexander01998 | All rights reserved.
+ * 
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/.
+ */
+package tk.wurst_client.v1_7_keybind_converter.menu;
+
+import javax.swing.JLabel;
+
+public class DoneMenu extends Menu
+{
+	public DoneMenu()
+	{
+		super("Done!", null);
+		hasNextMenu = false;
+		add(new JLabel("<html>"
+			+ "<p>Your alt list has been successfully converted.</p>"));
+	}
+	
+	@Override
+	public void showNextMenu()
+	{	
+		
+	}
+}
